pyproject.toml
[project]
name = "affluences-helper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot[webhooks]>=22.5",
    "requests>=2.32.5",
]
